Collapsed model folders down into structs.go for their respective packages
