--- pkg/external/models/dalleImages.go
@@ -1,5 +1,0 @@
-package external
-
-type DalleImages struct {
-	URL string `json:"url"`
-}
