--- pkg/external/models/dalleResponse.go
@@ -1,6 +1,0 @@
-package external
-
-type DalleResponse struct {
-	Created int           `json:"created"`
-	Data    []DalleImages `json:"data"`
-}
