--- pkg/external/models/openAIDivinciChoice.go
@@ -1,7 +1,0 @@
-package external
-
-type OpenAIDivinciChoice struct {
-	Text   string `json:"text"`
-	Index  int    `json:"index"`
-	Reason string `json:"finish_reason"`
-}
