--- pkg/external/models/openAIDivinciResponse.go
@@ -1,8 +1,0 @@
-package external
-
-type OpenAIDivinciResponse struct {
-	ID      string         `json:"id"`
-	Object  string         `json:"object"`
-	Model   string         `json:"model"`
-	Choices []OpenAIDivinciChoice `json:"choices"`
-}
