--- pkg/external/models/openAIGPTChoice.go
@@ -1,8 +1,0 @@
-package external
-
-type OpenAIGPTChoice struct {
-	Index         int              `json:"index"`
-	Message       OpenAIGPTMessage `json:"message"`
-	Finish_Reason string           `json:"finish_reason"`
-	LogProbs      string           `json:"logprobs"`
-}
