--- pkg/external/models/openAIGPTMessage.go
@@ -1,6 +1,0 @@
-package external
-
-type OpenAIGPTMessage struct {
-	Role    string `json:"role"`
-	Content string `json:"content"`
-}
