--- pkg/external/models/openAIGPTUsage.go
@@ -1,7 +1,0 @@
-package external
-
-type OpenAIGPTUsage struct {
-	Prompt_Tokens     int `json:"prompt_tokens"`
-	Completion_Tokens int `json:"completion_tokens"`
-	Total_Tokens      int `json:"total_tokens"`
-}
