+++ pkg/external/openAIDalle.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	external "main/pkg/external/models"
 	"main/pkg/util"
 	"net/http"
 	"os"
@@ -44,7 +43,7 @@ func GetDalleResponse(prompt string) (discordgo.File, error) {
 	}
 
 	responseBuffer, _ := io.ReadAll(httpResponse.Body)
-	var openAIResponse external.DalleResponse
+	var openAIResponse DalleResponse
 	err = json.Unmarshal([]byte(string(responseBuffer)), &openAIResponse)
 	if err != nil {
 		return discordgo.File{}, errors.New("error Parsing Response")
