+++ pkg/external/openAIDavinci.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"io"
 
-	external "main/pkg/external/models"
 	"main/pkg/logging"
 	"main/pkg/util"
 
@@ -37,7 +36,7 @@ func GetOpenAIResponse(prompt string) string {
 	}
 
 	buf, _ := io.ReadAll(resp.Body)
-	rspOAI := external.OpenAIGPTResponse{}
+	rspOAI := OpenAIGPTResponse{}
 	// TODO: This could contain an error from OpenAI (rate limit, server issue, etc)
 	// need to add proper error handling
 	err = json.Unmarshal([]byte(string(buf)), &rspOAI)
