--- pkg/persistance/models/stats.go
@@ -1,14 +1,0 @@
-package persistance
-
-type UserStats struct {
-	UserId string `json:"UserId"`
-
-	InteractionCount int `json:"InteractionCount"`
-	ChatCount        int `json:"ChatCount"`
-	GoodBotCount     int `json:"GoodBotCount"`
-	BadBotCount      int `json:"BadBotCount"`
-
-	ImageCount int `json:"ImageCount"`
-
-	LootBoxCount int `json:"LootBoxCount"`
-}
