--- pkg/persistance/models/stocks.go
@@ -1,6 +1,0 @@
-package persistance
-
-type Stock struct {
-	StockTicker string  `json:"stockTicker"`
-	StockCount  float64 `json:"stockCount"`
-}
