--- pkg/persistance/models/userEventCount.go
@@ -1,5 +1,0 @@
-package persistance
-
-type UserEventCount struct {
-	Count int `json:"count"`
-}
