+++ pkg/structs.go
@@ -1,9 +1,9 @@
-package model
+package pkg
 
 import "time"
 
 type Base struct {
-	ID           string `gorm:"primarykey"`
+	ID           string
 	DateCreated  time.Time
 	DateModified time.Time
 	CreatedBy    string
