+++ pkg/util/structs.go
